modified prog02 batch 2

+++ batch_2/print_not_equal_when_numbers_are_not_the_same.py
@@ -1,6 +1,6 @@
 #input
-num1 = float(input("Enter first number: "))
-num2 = float(input("Enter second number: "))
+num1 = int(input("Enter first number: "))
+num2 = int(input("Enter second number: "))
 
 #result
 print("Not Equal") if num1 != num2 else print("Equal")
